+++ lib/bitmex.rb
@@ -3,17 +3,25 @@ require 'hashie'
 
 require 'bitmex/version'
 require 'bitmex/mash'
-require 'bitmex/user'
+
 require 'bitmex/base'
+require 'bitmex/chat'
 require 'bitmex/client'
 require 'bitmex/trade'
 require 'bitmex/stats'
 require 'bitmex/quote'
 require 'bitmex/position'
 require 'bitmex/order'
+require 'bitmex/instrument'
+require 'bitmex/user'
+require 'bitmex/apikey'
 
+# Bitmex module
 module Bitmex
   class Error < StandardError; end
+
+  TESTNET_HOST = 'testnet.bitmex.com'.freeze
+  MAINNET_HOST = 'www.bitmex.com'.freeze
 
   def self.signature(api_secret, verb, path, expires, params)
     params = '' if params.nil?