+++ CHANGELOG.md
@@ -6,10 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Chat, instrument, apikey resources
 ### Changed
 - Trade resource
 - Stats, settlement, schema, quote, position resources
 - Orderbook, order resources
+- Liquidation, leaderboard, insurance, execution, user, announcement
 
 ## [0.0.2] - 2019-01-22
 ### Added